# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  token: string;
  email: string;
  message: string;
}

export interface Hero {
  id: string;
  label: string;
  title: string;
  subtitle: string;
  ctaText: string;
  ctaUrl: string;
  imageUrl: string;
  updatedAt: string;
}

export interface HeroInput {
  label: string;
  title: string;
  subtitle: string;
  ctaText: string;
  ctaUrl: string;
  imageUrl: string;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  longDescription: string;
  imageUrl: string;
  order: number;
  ctaText: string;
  ctaUrl: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  longDescription: string;
  imageUrl: string;
  order: number;
  ctaText: string;
  ctaUrl: string;
}

export interface Solution {
  id: string;
  name: string;
  description: string;
  imageUrl: string;
  order: number;
  updatedAt: string;
}

export interface SolutionInput {
  name: string;
  description: string;
  imageUrl: string;
  order: number;
}

export interface Section {
  id: string;
  name: string;
  title: string;
  description: string;
  imageUrl: string;
  videoUrl: string;
  order: number;
  updatedAt: string;
}

export interface SectionInput {
  name: string;
  title: string;
  description: string;
  imageUrl: string;
  videoUrl: string;
  order: number;
}

export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface Media {
  id: string;
  name: string;
  type: MediaType;
  url: string;
  size: string;
  createdAt: string;
}

export type MediaInputType = typeof MediaInputType[keyof typeof MediaInputType];


export const MediaInputType = {
  image: 'image',
  video: 'video',
} as const;

export interface MediaInput {
  name: string;
  type: MediaInputType;
}

export interface DeleteMediaInput {
  id: string;
}

export interface ContentSummary {
  products: number;
  solutions: number;
  sections: number;
  media: number;
  lastUpdated: string;
}

export interface ContentBundle {
  hero: Hero;
  products: Product[];
  solutions: Solution[];
  sections: Section[];
  media: Media[];
}