# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Mock admin authentication
  - name: content
    description: VOXA marketing content
  - name: media
    description: Mock media library
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Sign in to the admin dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: registerAdmin
      tags: [auth]
      summary: Register a mock admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "201":
          description: Account registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logoutAdmin
      tags: [auth]
      summary: Sign out
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /content/summary:
    get:
      operationId: getContentSummary
      tags: [content]
      summary: Get dashboard content summary
      responses:
        "200":
          description: Content summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentSummary"
  /content:
    get:
      operationId: getAllContent
      tags: [content]
      summary: Get all editable content
      responses:
        "200":
          description: All content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentBundle"
  /content/hero:
    get:
      operationId: getHero
      tags: [content]
      summary: Get hero content
      responses:
        "200":
          description: Hero content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hero"
    post:
      operationId: updateHero
      tags: [content]
      summary: Update hero content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeroInput"
      responses:
        "200":
          description: Updated hero
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hero"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content/products:
    get:
      operationId: getProducts
      tags: [content]
      summary: List products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [content]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content/products/{id}:
    get:
      operationId: getProduct
      tags: [content]
      summary: Get a product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [content]
      summary: Update product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [content]
      summary: Delete product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deleted product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content/solutions:
    get:
      operationId: getSolutions
      tags: [content]
      summary: List solutions
      responses:
        "200":
          description: Solution list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Solution"
    post:
      operationId: createSolution
      tags: [content]
      summary: Create solution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionInput"
      responses:
        "201":
          description: Created solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
  /content/solutions/{id}:
    get:
      operationId: getSolution
      tags: [content]
      summary: Get a solution
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSolution
      tags: [content]
      summary: Update solution
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionInput"
      responses:
        "200":
          description: Updated solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
    delete:
      operationId: deleteSolution
      tags: [content]
      summary: Delete solution
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deleted solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /content/sections:
    get:
      operationId: getSections
      tags: [content]
      summary: List page sections
      responses:
        "200":
          description: Section list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: createSection
      tags: [content]
      summary: Create page section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /content/sections/{id}:
    get:
      operationId: getSection
      tags: [content]
      summary: Get a page section
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSection
      tags: [content]
      summary: Update page section
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    delete:
      operationId: deleteSection
      tags: [content]
      summary: Delete section
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deleted section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /upload/media:
    get:
      operationId: getMedia
      tags: [media]
      summary: List media assets
      responses:
        "200":
          description: Media library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Create a mock media asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Uploaded asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
  /upload/delete:
    post:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteMediaInput"
      responses:
        "200":
          description: Deleted asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, email, message]
      properties:
        token:
          type: string
        email:
          type: string
        message:
          type: string
    Hero:
      type: object
      required: [id, label, title, subtitle, ctaText, ctaUrl, imageUrl, updatedAt]
      properties:
        id: { type: string }
        label: { type: string }
        title: { type: string }
        subtitle: { type: string }
        ctaText: { type: string }
        ctaUrl: { type: string }
        imageUrl: { type: string }
        updatedAt: { type: string }
    HeroInput:
      type: object
      required: [label, title, subtitle, ctaText, ctaUrl, imageUrl]
      properties:
        label: { type: string }
        title: { type: string }
        subtitle: { type: string }
        ctaText: { type: string }
        ctaUrl: { type: string }
        imageUrl: { type: string }
    Product:
      type: object
      required: [id, name, description, longDescription, imageUrl, order, ctaText, ctaUrl, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        longDescription: { type: string }
        imageUrl: { type: string }
        order: { type: number }
        ctaText: { type: string }
        ctaUrl: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [name, description, longDescription, imageUrl, order, ctaText, ctaUrl]
      properties:
        name: { type: string }
        description: { type: string }
        longDescription: { type: string }
        imageUrl: { type: string }
        order: { type: number }
        ctaText: { type: string }
        ctaUrl: { type: string }
    Solution:
      type: object
      required: [id, name, description, imageUrl, order, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        order: { type: number }
        updatedAt: { type: string }
    SolutionInput:
      type: object
      required: [name, description, imageUrl, order]
      properties:
        name: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        order: { type: number }
    Section:
      type: object
      required: [id, name, title, description, imageUrl, videoUrl, order, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        videoUrl: { type: string }
        order: { type: number }
        updatedAt: { type: string }
    SectionInput:
      type: object
      required: [name, title, description, imageUrl, videoUrl, order]
      properties:
        name: { type: string }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        videoUrl: { type: string }
        order: { type: number }
    Media:
      type: object
      required: [id, name, type, url, size, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [image, video] }
        url: { type: string }
        size: { type: string }
        createdAt: { type: string }
    MediaInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string, enum: [image, video] }
    DeleteMediaInput:
      type: object
      required: [id]
      properties:
        id: { type: string }
    ContentSummary:
      type: object
      required: [products, solutions, sections, media, lastUpdated]
      properties:
        products: { type: number }
        solutions: { type: number }
        sections: { type: number }
        media: { type: number }
        lastUpdated: { type: string }
    ContentBundle:
      type: object
      required: [hero, products, solutions, sections, media]
      properties:
        hero: { $ref: "#/components/schemas/Hero" }
        products:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        solutions:
          type: array
          items: { $ref: "#/components/schemas/Solution" }
        sections:
          type: array
          items: { $ref: "#/components/schemas/Section" }
        media:
          type: array
          items: { $ref: "#/components/schemas/Media" }