# lib/api-zod/src/generated/types/contentBundle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Hero } from './hero';
import type { Media } from './media';
import type { Product } from './product';
import type { Section } from './section';
import type { Solution } from './solution';

export interface ContentBundle {
  hero: Hero;
  products: Product[];
  solutions: Solution[];
  sections: Section[];
  media: Media[];
}