# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaType } from './mediaType';

export interface Media {
  id: string;
  name: string;
  type: MediaType;
  url: string;
  size: string;
  createdAt: string;
}