# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  description: string;
  longDescription: string;
  imageUrl: string;
  order: number;
  ctaText: string;
  ctaUrl: string;
  updatedAt: string;
}