# lib/api-zod/src/generated/types/sectionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SectionInput {
  name: string;
  title: string;
  description: string;
  imageUrl: string;
  videoUrl: string;
  order: number;
}