# lib/api-zod/src/generated/types/solution.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Solution {
  id: string;
  name: string;
  description: string;
  imageUrl: string;
  order: number;
  updatedAt: string;
}