# lib/api-zod/src/generated/types/solutionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SolutionInput {
  name: string;
  description: string;
  imageUrl: string;
  order: number;
}